+++ spec/netsuite/records/item_vendor_spec.rb
@@ -5,11 +5,23 @@ describe NetSuite::Records::ItemVendor do
 
   it 'has all the right fields' do
     [
-      :vendor, :purchase_price, :preferred_vendor
+      :purchase_price,
+      :preferred_vendor,
+      :vendor_code,
+      :vendor_currency_name,
     ].each do |field|
       expect(vendor).to have_field(field)
     end
   end
 
-  it 'has all the right record refs'
+  it 'has all the right record refs' do
+    [
+      :schedule,
+      :subsidiary,
+      :vendor,
+      :vendor_currency,
+    ].each do |field|
+      expect(vendor).to have_record_ref(field)
+    end
+  end
 end